+++ prism/src/hybrid/PH_StochTransient.cc
@@ -94,7 +94,7 @@ jdouble time		// time bound
 	double time_taken, time_for_setup, time_for_iters;
 	// misc
 	bool done;
-	int i, iters, num_iters;
+	long i, iters, num_iters;
 	double kb, kbt, max_diag, weight, term_crit_param_unif;
 	
 	// exception handling around whole function
