+++ prism/src/mtbdd/PM_StochBoundedUntil.cc
@@ -72,7 +72,7 @@ jlong __jlongpointer mu	// probs for multiplying
 	long start1, start2, start3, stop;
 	double time_taken, time_for_setup, time_for_iters;
 	// misc
-	int i, iters, num_iters;
+	long i, iters, num_iters;
 	double x, max_diag, weight, unif, term_crit_param_unif;
 	bool done, combine;
